test/AoC2021.Tests/Days/Day20Tests.cs
using System.IO;
using System.Reflection;
using Xunit;
using AoC2021.Days;

namespace AoC2021.Tests.Days
{
    public class Day20Tests
    {
        [Fact]
        public void Day20Test1()
        {
            // arrange
            var path = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
            path = Path.GetFullPath(Path.Combine(path, "..", "..", ".."));

            var inputFile = Directory.GetFiles(path + @"/TestInput", "Day20_1.txt")[0];
            
            // on day 20 the sample input is fundamentally different
            // from the actual input. I wrote the solution with
            // assumptions about the input, and so it works on the actual
            // input but not on the sample. Skip this test.

            // var day20 = new Day20(inputFile);

            // act
            // var result1 = day20.PartOne();
            // var result2 = day20.PartTwo();

            // assert
            // Assert.Equal("35", result1);
            // Assert.Equal("3351", result2);
        }
    }
}